Added basic API to MetroClient to create contextMenu buttons
The DataSource passes a title and a functionName to the MetroClient's
API function createContextMenuButton(), causing a button to be made.
The MetroClient keeps track of state to make sure that duplicate
buttons are not made every time a page loads.

When the button is pressed, the MetroClient will call the relevant
function.