Adds the basic wiring for a modal input box.
When a metroClient function is called with a description and callback
function, a HTML div is loaded by the background script and displayed on
the page. The div contains a form with one text input and the
description for the input and when the submit button is hit the callback
from the datasource is called.

TODO:
  - Make the modal not be just a plain HTML mess at the bottom of the
page and look like an actual dialog box.
  - Finish off the datasource to show this functionality.