Improve DS loading system, refactor metroClient and 'DS' var
Previously we couldn't load multiple DataSources due to namespace
conflicts. DataSources are now expected to be namespaced in an object
and must contain a 'slug' property which we can use for identification.
To launch itself, a DataSource must pass its reference to a
registerDataSource() function which Metro provides.

A devMode flag is passed when loading DataSources, so we know how to
identify DataSources. This should be improved so that it's not necessary
- how?

I moved the MetroClient instantiation into its own function for clarity,
since it's huge. Maybe we should put it into a separate file along with
its util functions?

I also changed the DS variable to be 'slug', because that's what it
actually is.

Lastly, I fixed the params passed to the dev-mode loadSourceFromBaseURL
function in the scriptLoader