Adds counter disable btn, moves it to shadow dom
The counter is implemented in a Shadow DOM now, to protect it from the
page's CSS.

The disable button allows users to disable the counter if they find it
annoying.

I also fixed the checkboxes in the extension so that they assume sane
defaults upon installing the extension. Previously, they didn't work
because the if-statement was checking the value of the setting and not
checking its existence.